refactor(refactor/tier): discovery 도메인을 추가하고 관련 로직을 전부 이동시킴
