+++ src/main/java/com/kustaurant/kustaurant/common/discovery/domain/RestaurantTierMapDTO.java
@@ -1,6 +1,5 @@
-package com.kustaurant.kustaurant.common.restaurant.domain.dto;
+package com.kustaurant.kustaurant.common.discovery.domain;
 
-import com.kustaurant.kustaurant.common.restaurant.constants.MapConstants;
 import com.kustaurant.kustaurant.common.restaurant.domain.Coordinate;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
