+++ src/main/java/com/kustaurant/kustaurant/common/discovery/infrastructure/RestaurantDiscoveryJpaRepository.java
@@ -1,0 +1,14 @@
+package com.kustaurant.kustaurant.common.discovery.infrastructure;
+
+import com.kustaurant.kustaurant.common.restaurant.infrastructure.restaurant.RestaurantEntity;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface RestaurantDiscoveryJpaRepository extends JpaRepository<RestaurantEntity, Integer> {
+    List<RestaurantEntity> findAll(Specification<RestaurantEntity> spec);
+    Page<RestaurantEntity> findAll(Specification<RestaurantEntity> spec, Pageable pageable);
+}
