+++ src/main/java/com/kustaurant/kustaurant/common/restaurant/argument_resolver/LocationListArgumentResolver.java
@@ -1,6 +1,6 @@
 package com.kustaurant.kustaurant.common.restaurant.argument_resolver;
 
-import com.kustaurant.kustaurant.common.restaurant.domain.enums.LocationEnum;
+import com.kustaurant.kustaurant.common.discovery.enums.LocationEnum;
 import com.kustaurant.kustaurant.global.exception.exception.ParamException;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.core.MethodParameter;
