+++ src/main/java/com/kustaurant/kustaurant/common/restaurant/service/port/RestaurantRepository.java
@@ -17,9 +17,6 @@ public interface RestaurantRepository {
     List<Restaurant> findByCuisineAndPositionAndStatus(String cuisine, String position, String status);
 
     Restaurant save(Restaurant restaurant);
-    // 검색결과 페이징
-    List<Restaurant> findAll(Specification<RestaurantEntity> spec);
-    List<Restaurant> findAll(Specification<RestaurantEntity> spec, Pageable pageable);
 
     // TODO: need to delete everything below this
     List<RestaurantEntity> findAll();
