+++ src/main/resources/templates/restaurantComments.html
@@ -19,7 +19,7 @@
                 <span th:text="${restaurantComment.commentScore}"></span>
             </div>
             <div class="nick-date-div">
-                <img th:src="${restaurantComment.user.getRankImg()}">
+                <img th:src="${restaurantComment.userEntity.getRankImg()}">
                 <span class="nick-span" th:text="${restaurantComment.commentNickname}"></span>
                 <span class="date-span" th:text="${restaurantComment.commentTime}"></span>
                 <!-- Button trigger modal -->
@@ -52,7 +52,7 @@
             </div>
             <div class="body-div">
                 <div class="nick-date-div">
-                    <img th:src="${reply.user.getRankImg()}">
+                    <img th:src="${reply.userEntity.getRankImg()}">
                     <span class="nick-span" th:text="${reply.commentNickname}"></span>
                     <span class="date-span" th:text="${reply.commentTime}"></span>
                     <!-- Button trigger modal -->
